pyproject.toml
[project]
name = "lithium"
version = "0.1.0"
description = "Starter template for making a custom Reasoning and Action agent (using tool calling) in LangGraph."
readme = "README.md"
requires-python = ">=3.11,<=4.0"

authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
    { name = "Ben Lin", email = "blin7.webdev@gmail.com" },
]
dependencies = [
    "langchain-anthropic>=1.4.0",
    "langchain>=1.2.15",
    "langchain-fireworks>=0.1.7",
    "python-dotenv>=1.2.2",
    "langchain-tavily>=0.2.17",
    "pydantic-settings>=2.13.1",
    "xml-pydantic>=0.1.0",
    "fastapi>=0.135.3",
    "uvicorn>=0.44.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary,pool]>=3.1.0"
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.15.1", "ipykernel"]

[tool.uv.sources]
xml-pydantic = { path = "packages/xml-pydantic", editable = true }

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["orient"]
[tool.setuptools.package-dir]
"orient" = "app/orient"


[tool.setuptools.package-data]
"*" = ["py.typed"]


[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "deepdiff>=8.6.2",
    "ipykernel>=7.2.0",
    "langgraph-cli[inmem]>=0.4.19",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
        "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "T201",
    "UP",# pyupgrade
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501", # line length (handled by formatter)
    # FastAPI's Depends() in default args is intentional, not a bug.
    "B008",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]  # test names are the documentation

[tool.pytest.ini_options]
pythonpath = [".", "app"]
testpaths = ["tests/unit_tests/"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-config=.coveragerc",
    "--cov-report=term-missing",
]
[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.pydocstyle]
convention = "google"

